refactored treecell into package
