+++ plep_main/src/deltadak/ui/taskcell/ChangeListenerWithBlocker.java
@@ -1,4 +1,4 @@
-package deltadak.ui;
+package deltadak.ui.taskcell;
 
 import javafx.beans.value.ChangeListener;
 import javafx.beans.value.ObservableValue;
